requirements.txt
google-genai
python-dotenv==1.2.1
protobuf==6.33.5
